+++ src/main/scala/sangria/introspection/package.scala
@@ -152,7 +152,7 @@ package object introspection {
           }
         }),
       Field("interfaces", OptionType(ListType(__Type)), resolve = _.value._2 match {
-        case t: ObjectType[_, _] ⇒ Some(t.interfaces.asInstanceOf[List[Type]] map (true → _))
+        case t: ObjectType[_, _] ⇒ Some(t.allInterfaces.asInstanceOf[Vector[Type]] map (true → _))
         case _ ⇒ None
       }),
       Field("possibleTypes", OptionType(ListType(__Type)), resolve = ctx ⇒ ctx.value._2 match {